Moved copy to clipboard logic to activity item
