+++ source/ui/ActivityItem/components/items/TokenItem.js
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React from 'react';
 import NumberFormat from 'react-number-format';
 import moment from 'moment';
 import PropTypes from 'prop-types';
@@ -6,25 +6,12 @@ import { useTranslation } from 'react-i18next';
 import { capitalize } from '@material-ui/core';
 
 import UnknownIcon from '@assets/icons/unknown-icon.svg';
-import shortAddress from '@shared/utils/short-address';
 import { GenericIcon } from '@ui';
 
 import ActivityItemDisplay from '../ActivityItemDisplay';
 import ActivityItemDetails from '../ActivityItemDetails';
 
-const getSubtitle = (type, to, from, t) => (({
-  SEND: ` · ${t('activity.subtitle.to')}: ${shortAddress(to)}`,
-  BURN: ` · ${t('activity.subtitle.to')}: ${shortAddress(to)}`,
-  RECEIVE: ` · ${t('activity.subtitle.from')}: ${shortAddress(from)}`,
-})[type]);
-
-const getAddress = (type, to, from, canisterId) => (
-  {
-    SEND: to,
-    BURN: to,
-    RECEIVE: from,
-  }
-)[type] || canisterId || '';
+import { getSubtitle } from '../../utils';
 
 const TokenItem = (props) => {
   const {
@@ -41,32 +28,12 @@ const TokenItem = (props) => {
     setOpenDetail,
     isTransaction,
     hovering,
+    copied,
+    onCopy,
+    tooltipText,
   } = props;
   const { t } = useTranslation();
-  const [copied, setCopied] = useState(false);
-  const copyText = t('copy.copyTextAddress');
-  const copiedText = t('copy.copiedText');
-  const [tooltipText, setTooltipText] = useState(copyText);
 
-  const handleClickCopy = (e) => {
-    e.stopPropagation();
-
-    /* eslint-disable no-nested-ternary */
-    navigator.clipboard.writeText(
-      getAddress(type, to, from, canisterId),
-    );
-
-    setCopied(true);
-    setTooltipText(copiedText);
-
-    setTimeout(() => {
-      setCopied(false);
-    }, 1000);
-
-    setTimeout(() => {
-      setTooltipText(copyText);
-    }, 1500);
-  };
   return (
     <>
       <ActivityItemDisplay
@@ -81,7 +48,7 @@ const TokenItem = (props) => {
         tooltip={getSubtitle(type, to, from, t, canisterId)}
         copied={copied}
         tooltipText={tooltipText}
-        onCopy={handleClickCopy}
+        onCopy={onCopy}
       />
       <ActivityItemDetails
         main={<NumberFormat value={amount} displayType="text" thousandSeparator="," suffix={` ${symbol}`} decimalScale={5} />}
@@ -128,4 +95,7 @@ TokenItem.propTypes = {
   setOpenDetail: PropTypes.func.isRequired,
   isTransaction: PropTypes.bool.isRequired,
   hovering: PropTypes.bool,
+  copied: PropTypes.bool.isRequired,
+  onCopy: PropTypes.func.isRequired,
+  tooltipText: PropTypes.string.isRequired,
 };
