+++ source/ui/ActivityItem/utils.js
@@ -1,0 +1,15 @@
+import shortAddress from '@shared/utils/short-address';
+
+export const getSubtitle = (type, to, from, t) => (({
+  SEND: ` · ${t('activity.subtitle.to')}: ${shortAddress(to)}`,
+  BURN: ` · ${t('activity.subtitle.to')}: ${shortAddress(to)}`,
+  RECEIVE: ` · ${t('activity.subtitle.from')}: ${shortAddress(from)}`,
+})[type]);
+
+export const getAddress = (type, to, from, canisterId) => (
+  {
+    SEND: to,
+    BURN: to,
+    RECEIVE: from,
+  }
+)[type] || canisterId || '';
